requirements.txt
appdirs==1.4.0
beautifulsoup4==4.5.3
packaging==16.8
pycrypto==2.6.1
pyparsing==2.1.10
requests==2.13.0
six==1.10.0
